Update deque interface.
The deque interface is updated to be a bit cleaner and more restrictive.
Nodes cannot be created except by insertion to a deque. Insertion is
only allowed at the head or tail of the deque. There is now only one
kind of deletion (remove link and destroy), rather than 3 as before
(remove link, destroy link, and remove and destroy link).