Mostly working.
Seems to be running the files without errors. Biggest change: now the
sched_queue update and the register file update both happen in the same
cycle, at the beginning of the cycle, the cycle after the instruction
completes. Not sure, but I think this is what is being asked in the
project description."